removing redundant code, refactoring models dir
