+++ src/teehr/models/loading/nwm30_grid.py
@@ -3,142 +3,13 @@ from typing import Optional
 
 from pydantic import BaseModel
 
-
-# GRID DATA VARIABLES ENUMS
-class LandAssimVariablesEnum(str, Enum):
-    ACCET = "ACCET"
-    ACSNOM = "ACSNOM"
-    EDIR = "EDIR"
-    FSNO = "FSNO"
-    ISNOW = "ISNOW"
-    QRAIN = "QRAIN"
-    QSNOW = "QSNOW"
-    SNEQV = "SNEQV"
-    SNLIQ = "SNLIQ"
-    SNOWH = "SNOWH"
-    SNOWT_AVG = "SNOWT_AVG"
-    SOILICE = "SOILICE"
-    SOILSAT_TOP = "SOILSAT_TOP"
-    SOIL_M = "SOIL_M"
-    SOIL_T = "SOIL_T"
-
-
-class LandShortVariablesEnum(str, Enum):
-    ACCET = "ACCET"
-    SNOWT_AVG = "SNOWT_AVG"
-    SOILSAT_TOP = "SOILSAT_TOP"
-    FSNO = "FSNO"
-    SNOWH = "SNOWH"
-    SNEQV = "SNEQV"
-
-
-class LandMediumVariablesEnum(str, Enum):
-    FSA = "FSA"
-    FIRA = "FIRA"
-    GRDFLX = "GRDFLX"
-    HFX = "HFX"
-    LH = "LH"
-    UGDRNOFF = "UGDRNOFF"
-    ACCECAN = "ACCECAN"
-    ACCEDIR = "ACCEDIR"
-    ACCETRAN = "ACCETRAN"
-    TRAD = "TRAD"
-    SNLIQ = "SNLIQ"
-    SOIL_T = "SOIL_T"
-    SOIL_M = "SOIL_M"
-    SNOWH = "SNOWH"
-    SNEQV = "SNEQV"
-    ISNOW = "ISNOW"
-    FSNO = "FSNO"
-    ACSNOM = "ACSNOM"
-    ACCET = "ACCET"
-    CANWAT = "CANWAT"
-    SOILICE = "SOILICE"
-    SOILSAT_TOP = "SOILSAT_TOP"
-    SNOWT_AVG = "SNOWT_AVG"
-
-
-class LandLongVariablesEnum(str, Enum):
-    UGDRNOFF = "UGDRNOFF"
-    SFCRNOFF = "SFCRNOFF"
-    SNEQV = "SNEQV"
-    ACSNOM = "ACSNOM"
-    ACCET = "ACCET"
-    CANWAT = "CANWAT"
-    SOILSAT_TOP = "SOILSAT_TOP"
-    SOILSAT = "SOILSAT"
-
-
-class ForcingVariablesEnum(str, Enum):
-    U2D = "U2D"
-    V2D = "V2D"
-    T2D = "T2D"
-    Q2D = "Q2D"
-    LWDOWN = "LWDOWN"
-    SWDOWN = "SWDOWN"
-    RAINRATE = "RAINRATE"
-    PSFC = "PSFC"
-
-
-# OUTPUT TYPE ENUMS: Potential output type names
-class ShortAndAnalysisOutputEnum(str, Enum):
-    land = "land"
-
-
-class MediumOutputEnum(str, Enum):
-    land_1 = "land_1"
-    land_2 = "land_2"
-    land_3 = "land_3"
-    land_4 = "land_4"
-    land_5 = "land_5"
-    land_6 = "land_6"
-    land_7 = "land_7"
-
-
-class LongOutputEnum(str, Enum):
-    land_1 = "land_1"
-    land_2 = "land_2"
-    land_3 = "land_3"
-    land_4 = "land_4"
-
-
-class ForcingOutputEnum(str, Enum):
-    forcing = "forcing"
-
-
-# OUTPUT TYPE MODELS
-class Analysis(BaseModel):
-    output_type: ShortAndAnalysisOutputEnum
-    land: Optional[LandAssimVariablesEnum]
-
-
-class ShortRange(BaseModel):
-    output_type: ShortAndAnalysisOutputEnum
-    land: Optional[LandShortVariablesEnum]
-
-
-class MediumRange(BaseModel):
-    output_type: MediumOutputEnum
-    land_1: Optional[LandMediumVariablesEnum]
-    land_2: Optional[LandMediumVariablesEnum]
-    land_3: Optional[LandMediumVariablesEnum]
-    land_4: Optional[LandMediumVariablesEnum]
-    land_5: Optional[LandMediumVariablesEnum]
-    land_6: Optional[LandMediumVariablesEnum]
-    land_7: Optional[LandMediumVariablesEnum]
-
-
-class LongRange(BaseModel):
-    output_type: LongOutputEnum
-    land_1: Optional[LandLongVariablesEnum]
-    land_2: Optional[LandLongVariablesEnum]
-    land_3: Optional[LandLongVariablesEnum]
-    land_4: Optional[LandLongVariablesEnum]
-
-
-class Forcing(BaseModel):
-    output_type: ForcingOutputEnum
-    forcing: Optional[ForcingVariablesEnum]
+from teehr.models.loading.nwm22_grid import (
+    Analysis,
+    ShortRange,
+    MediumRange,
+    LongRange,
+    Forcing,
+)
 
 
 # CONFIGURATIONS ENUM: All possible configuration
