+++ tests/test_nwm22_grid_loading.py
@@ -2,6 +2,7 @@
 from pathlib import Path
 import filecmp
 
+import pandas as pd
 import pytest
 
 from teehr.loading.nwm22.nwm_grid_data import fetch_and_format_nwm_grids
@@ -14,7 +15,9 @@ WEIGHTS_FILEPATH = Path(TEST_DIR, "onehuc10_weights.parquet")
 @pytest.mark.filterwarnings("ignore::RuntimeWarning")
 def test_grid_loading():
 
-    component_paths = ["gcs://national-water-model/nwm.20201218/forcing_analysis_assim/nwm.t00z.analysis_assim.forcing.tm00.conus.nc"] # noqa
+    component_paths = [
+        "gcs://national-water-model/nwm.20201218/forcing_analysis_assim/nwm.t00z.analysis_assim.forcing.tm00.conus.nc" # noqa
+    ]
 
     json_paths = build_zarr_references(component_paths,
                                        TEST_DIR,
@@ -36,7 +39,11 @@ def test_grid_loading():
 
     parquet_file = Path(TEST_DIR, "20201218T00Z.parquet")
     test_file = Path(TEST_DIR, "grid_benchmark.parquet")
-    assert filecmp.cmp(test_file, parquet_file, shallow=False)
+
+    bench_df = pd.read_parquet(test_file)
+    test_df = pd.read_parquet(parquet_file)
+
+    assert test_df.compare(bench_df).index.size == 0
 
 
 if __name__ == "__main__":
